+++ enclone_tail/src/lib.rs
@@ -9,5 +9,6 @@ pub mod neighbor;
 pub mod newick;
 pub mod plot;
 pub mod polygon;
+pub mod print_stats;
 pub mod string_width;
 pub mod tail;
